Add API to get variable axis information from font
This adds QFontInfo::variableAxes() which can be used to inspect
a variable font and see which axes it defines. It is implemented
on DirectWrite, CoreText and FreeType.

On FreeType, there is a bug where getting the information about
variable axes on a font will cause the FT_Face to be unusable.
Unfortunately this happens for Cantarell-VF which is the default
font on some platforms. To work around this, we create a
temporary duplicate of the FT_Face for multiple masters fonts
on older FreeType versions when we fetch the information about
the axes so that we can throw away the broken FT_Face after.

[ChangeLog][QtGui][Fonts] Added API to get variable axis
information for a font.

Change-Id: Ia16e0a5f6df5744dc47e7f16c074391afde1f126
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>