Add data for timezone L10n and enable the feature to activate it
In the process, fix the timezone_locale feature's LABEL to distinguish
it from feature timezone. Remove various QEXPECT_FAIL()s for tests
that now (finally) work as they should thanks to the new data. Improve
the reporting from some of those tests.

This commit is separated from its parent because the data is huge: at
12 MB, it causes problems for browsers displaying the full data. The
parent contains stubs for the relevant tables that document the size
of the data and what it compiles down to: this commit replaces those
with the real thing.

Task-number: QTBUG-32759
Fixes: QTBUG-115158
Change-Id: I46f9fa167c82236315b606f8e0e43a90d63dd3eb
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>