Android: Prevent Qt apps from running in the background
Previously, the environment variable QT_BLOCK_EVENT_LOOPS_WHEN_SUSPENDED
was set to an empty string, which allowed QAndroidEventDispatcherStopper
to keep the application running in the background when it entered the
suspended state (onStop).
To address this issue, the isBackgroundRunningBlocked() function has
been added to validate the proper value when getMetaData(..) returns a
value.

Fixes: QTBUG-125496
Pick-to: 6.7 6.8
Change-Id: I541af3bb1f4e239ee8ba2be32583b91afc7a43e2
Reviewed-by: Tinja Paavoseppä <tinja.paavoseppa@qt.io>