a11y: Report actual platform window name on Windows
Remember the value set for the window name
in QWindowsWindow::setWindowTitle.

Override QPlatformWindow::windowTitle introduced
in a previous commit for QWindowsWindow and let
it return the formatted window name as it
was actually set.

With this in place, the NVDA screen reader on
Windows now announces "MyApp, window" as expected
when the sample app from QTBUG-127563 receives focus, and
the accessible name that is now matching the displayed window
title can also be seen in NVDA's interactive Python
console:

    >>> focus.name
    'MyApp'

Fixes: QTBUG-127563
Change-Id: Ifca165d6c302199d771ce0805c4d2d91047ca31c
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>