Fix #if-ery for MS-specific tst_QDateTime exceptions
The issues with its timezone data are only relevant when we're
actually using the backend based on its registry. So introduce the
same USING_WIN_TZ define as tst_QTimeZone uses and apply it to some
exceptions previously testing Q_OS_WIN. For clarity, also use it for
the second pace feature ICU was tested.

Change-Id: Ibf88bc72260dda6eae3317b3ff3117670e6531b2
Reviewed-by: Øystein Heskestad <oystein.heskestad@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>