Q(Multi)Map: update (in)equality operators
Use comparison helper macros for the inequality operators.
Note that we cannot use the "public" macros, because we need to provide
attributes for the comparison of the same type, so use the _HELPER
macro.

Extend the unit tests with the compile-time checks + provide the
missing tests for QMultiMap (in)equality operators.

Task-number: QTBUG-120305
Change-Id: I0a622cea29079b8437bebc873b7ba97986516eed
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>