Q(Multi)Hash: update (in)equality operators
Use comparison helper macros for the inequality operators.
Note that we cannot use the "public" macros, because we need to provide
attributes for the comparison of the same type, so use the _HELPER
macro.

Convert the operators to hidden friends. We do not need to use
QT_CORE_REMOVED_SINCE in this case, because the classes are templates,
so the operators were not exported.

Add the missing docs for the (in)equality operators of QMultiHash.

gcc-10 complains about accessing const_iterator::i from a hidden
friend function, claiming that it is private in that context:

 qtbase/src/corelib/tools/qhash.h:916:38: error:
  ‘QHash<QString, QVariant>::const_iterator::piter
   QHash<QString, QVariant>::const_iterator::i’ is private within
   this context
 if (i == lhs.end() || !i.i.node()->valuesEqual(it.i.node()))
                        ~~^

Use a static helper method that does this comparison.

Task-number: QTBUG-120305
Change-Id: I41b8889b659ae5dd858790a4ff629807e6666dc4
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>