QSet: update (in)equality operators
Use comparison helper macros for the (in)equality operators.
Note that we cannot use the "public" macros, because we need to provide
attributes for the comparison of the same type, so use the _HELPER
macro instead.

Convert the operators to hidden friends. We do not need to use
QT_CORE_REMOVED_SINCE in this case, because the class is a template,
so the operators were not exported.

Task-number: QTBUG-120305
Change-Id: I575ab953c4693b6550cd764919671df8dfa5624a
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>