wasm: Add the qdir test
There is a problem in that all the tests share the same resource system,
to not interfere with other tests, the resource files are placed in
its own subdirectory.

Also, wrt long filenames, it seems wasm truncates them but does not
fail. This is visible in the long filename test.

There are two bugs created due to test failures:
QTBUG-127767
QTBUG-127766

Change-Id: Ie3c08b4a105f1122284d1f99677597cc37dc91a7
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>