rhi: gl: Use GL_EXT_multisampled_render_to_texture, if present, always
Meaning on the common MSAA path that involves passing in a multisample
renderbuffer and a non-multisample resolve texture, i.e. the traditional
GLES 3.0 compatible approach, that is used throughout Qt in Qt Quick and
elsewhere.

When EXT_multisampled_render_to_texture (not to be confused with
non-GLES extensions or similarly named non-GLES, core functions), is
present, which is relatively rare in practice (probably offered on Mali
and Qualcomm perhaps), we could ignore the multisample renderbuffer
object completely, and get automatic resolving into the
(non-multisample) texture.

This is already done on the true texture-based path, mainly because
using the sibling extension for multiview is mandatory when dealing with
multiview (texture arrays) and MSAA on e.g. the Quest 3. However, the
texture-based code path is not generally utilized in the rest of Qt when
doing MSAA since renderbuffer+explicit resolve into a texture is GLES
3.0 compatible, whereas multisample textures (and then resolve into a
texture) needs GLES 3.1.

So

1. use glFramebufferTexture2DMultisampleEXT instead of
glFramebufferRenderbuffer when the conditions are met (multisample
renderbuffer, there is a resolve texture specified, and the extension is
supported), and

2. skip the explicit resolve (glBlitFramebuffer) when the conditions for
1) were met.

Do this only for color attachments. As the docs for
setDepthResolveTexture() state, we do not actually support
depth(stencil) resolving with OpenGL ES in combination with
QRhiRenderBuffer, so do not have to worry about that here.

Change-Id: I0bf88089c3b1b15a8ccc5652e1f918a2dad215e6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>